% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_functions.R
\name{arma_to_ma}
\alias{arma_to_ma}
\title{Transform an an ARMA(p, q) into a truncated infinite-order MA representation}
\usage{
arma_to_ma(model, max_lag = 30)
}
\arguments{
\item{model}{A list containing the output ARMA model from function \code{gen_arma()}}

\item{max_lag}{Integer. Number of MA coefficients \eqn{\psi_j} to compute.
The output vector has length \code{max_lag + 1}, as \eqn{\psi_0 = 1}.}
}
\value{
A numeric vector of length \code{max_lag + 1} containing the MA coefficients
\eqn{\psi_0,\psi_1,\ldots,\psi_{max\_lag}} of the infinite-order MA representation.
}
\description{
Converts an ARMA(p, q) model into its infinite-order MA representation
by computing the first \code{max_lag} MA coefficients \eqn{\psi_j}.
The algorithm follows the method in
\emph{Introduction to Time Series and Forecasting} by Brockwell & Davis (3rd ed.).
}
\details{
The function computes the coefficients in the infinite moving-average expansion
of an ARMA(p, q) model:
\deqn{
  X_t = \sum_{j=0}^{\infty} \psi_j W_{t-j},
}
where:
\deqn{
  \psi_j = \theta_j + \sum_{k=1}^{\min(p, j)} \phi_k \psi_{j-k},
}
and \eqn{\theta_j = 0} for \eqn{j > q}.
}
\examples{
n <- 100
phi <- c(0.5, 0.2)
theta <- -0.3
wn <- rnorm(n)
model <- gen_arma(n, wn, ar_coefs = phi, ma_coefs = theta)

arma_to_ma(model, max_lag = 10)

}
