% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma.R
\name{fit_ARMA}
\alias{fit_ARMA}
\title{Fit an ARMA Model to Data}
\usage{
fit_ARMA(data, p_order = 0, q_order = 0)
}
\arguments{
\item{data}{Data frame containing at least \code{y} (and optionally \code{t}), or a
numeric response vector.}

\item{p_order}{Integer AR order.}

\item{q_order}{Integer MA order.}
}
\value{
An object of class Arima.
}
\description{
Fits an ARMA model of order \code{p_order} and \code{q_order}. If
\code{data} is a data frame with a \code{t} column, \code{t} is included as an external regressor.
}
\examples{
d <- generate_ARMA_dataset(n = 300, p = c(0.4), q = c(0.25, -0.1), sigma = 1.5, b = c(0.04, 2))
fit <- fit_ARMA(d, p_order = 1, q_order = 2)

}
