% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_functions.R
\name{get_theoretical_acf}
\alias{get_theoretical_acf}
\title{Computes the theoretical autocorrelation of an ARMA(p,q) model}
\usage{
get_theoretical_acf(model, sigma, max_lag = 30)
}
\arguments{
\item{model}{A list containing the output ARMA model from function \code{gen_arma()}}

\item{sigma}{Numeric. Standard deviation of the white noise process.}

\item{max_lag}{Integer. Maximum lag for which to compute the ACF.}
}
\value{
A numeric vector of length \code{max_lag + 1} containing:
\deqn{\rho(0), \rho(1), \ldots, \rho(\text{max\_lag}).}
}
\description{
Computes the theoretical autocorrelation function (ACF) of an
ARMA(p, q) model.
}
\details{
The ACF is obtained by normalizing the ACVF:
\deqn{
  \rho(j) = \frac{\gamma(j)}{\gamma(0)}.
}

Note: If the AR part is not causal, a warning is printed and the function
still returns the normalized partial ACF.
}
\examples{
n <- 100
phi <- c(0.5, 0.2)
theta <- -0.3
wn <- rnorm(n)
model <- gen_arma(n, wn, ar_coefs = phi, ma_coefs = theta)

get_theoretical_acf(model, sigma = 1, max_lag = 10)

}
