% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_functions.R
\name{is_causal}
\alias{is_causal}
\title{Determines whether an ARMA(p,q) model is causal}
\usage{
is_causal(model)
}
\arguments{
\item{model}{A list containing the output ARMA model from function \code{gen_arma()}}
}
\value{
Logical.
\code{TRUE} if the AR component is causal, \code{FALSE} otherwise.
}
\description{
Checks whether an ARMA model is causal. This can be determined by
whether its AR polynomial has all roots strictly outside the unit circle.
}
\details{
The AR characteristic polynomial is:
\deqn{
  1 - \phi_1 z - \cdots - \phi_p z^p = 0.
}
The process is causal if all roots satisfy \eqn{|z| > 1}.
If the model has no AR part (\code{ar_coefs = NULL}), the function returns TRUE.
}
\examples{
n <- 100
wn <- rnorm(n)
model1 <- gen_arma(n, wn, ar_coefs = 0.2, ma_coefs = NULL)
model2 <- gen_arma(n, wn, ar_coefs = 1.2, ma_coefs = NULL)

is_causal(model1) # TRUE
is_causal(model2) # FALSE

}
