% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma.R
\name{plot_ARMA_series}
\alias{plot_ARMA_series}
\title{Plot ARMA series over time}
\usage{
plot_ARMA_series(x, series = c("y", "e"), fitted_vals = NULL, fit = NULL)
}
\arguments{
\item{x}{Either a data frame returned by \code{generate_ARMA_dataset()} /
\code{generate_ARMA_errors()}, or a numeric vector.}

\item{series}{Character string. Either \code{"y"} or \code{"e"}.}

\item{fitted_vals}{Optional numeric vector of fitted values.}

\item{fit}{Optional fitted model object from \code{fit_ARMA()}.}
}
\value{
No return value; called for side effects.
}
\description{
Plots ARMA data or errors from either a generated ARMA data
frame or a numeric vector and overlays fitted values if provided.
}
\examples{
d <- generate_ARMA_dataset(n = 300, p = c(0.4), q = c(0.25, -0.1), sigma = 1.5, b = c(0.04, 2))
fit <- fit_ARMA(d, p_order = 1, q_order = 2)
plot_ARMA_series(d, series = "y")
plot_ARMA_series(d, series = "e", fit = fit)
plot_ARMA_series(d$y, series = "y")

}
